Change balance type to BigDecimal.

+++ src/main/java/net/lugburz/tinytransfer/repository/Account.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 /**
  * Minimal model for a bank account.
  */
@@ -12,5 +14,5 @@ import lombok.Getter;
 @EqualsAndHashCode
 public final class Account {
     private final String accountNo;
-    private final long balance;
+    private final BigDecimal balance;
 }
